Refactor global decorator
